Review suggestions
- Style changes,
- Allow not filling a list entirely in user-provided shader parameters,
- Fix the user-defined binding id,
